breaking out the classes
