+++ src/main/java/com/neuronrobotics/bowlerstudio/av/ImagesToVideo.java
@@ -315,6 +315,7 @@ public class ImagesToVideo implements ControllerListener, DataSinkListener {
 	/**
 	 * Create a media locator from the given string.
 	 */
+	@SuppressWarnings("unused")
 	static MediaLocator createMediaLocator(String url) {
 
 		MediaLocator ml;
@@ -339,176 +340,7 @@ public class ImagesToVideo implements ControllerListener, DataSinkListener {
 	// Inner classes.
 	///////////////////////////////////////////////
 
-	/**
-	 * A DataSource to read from a list of JPEG image files and turn that into a
-	 * stream of JMF buffers. The DataSource is not seekable or positionable.
-	 */
-	class ImageDataSource extends PullBufferDataSource {
 
-		ImageSourceStream streams[];
 
-		ImageDataSource(int width, int height, int frameRate, Vector images) {
-			streams = new ImageSourceStream[1];
-			streams[0] = new ImageSourceStream(width, height, frameRate, images);
-		}
 
-		public void setLocator(MediaLocator source) {
-		}
-
-		public MediaLocator getLocator() {
-			return null;
-		}
-
-		/**
-		 * Content type is of RAW since we are sending buffers of video frames
-		 * without a container format.
-		 */
-		public String getContentType() {
-			return ContentDescriptor.RAW;
-		}
-
-		public void connect() {
-		}
-
-		public void disconnect() {
-		}
-
-		public void start() {
-		}
-
-		public void stop() {
-		}
-
-		/**
-		 * Return the ImageSourceStreams.
-		 */
-		public PullBufferStream[] getStreams() {
-			return streams;
-		}
-
-		/**
-		 * We could have derived the duration from the number of frames and
-		 * frame rate. But for the purpose of this program, it's not necessary.
-		 */
-		public Time getDuration() {
-			return DURATION_UNKNOWN;
-		}
-
-		public Object[] getControls() {
-			return new Object[0];
-		}
-
-		public Object getControl(String type) {
-			return null;
-		}
-	}
-
-	/**
-	 * The source stream to go along with ImageDataSource.
-	 */
-	class ImageSourceStream implements PullBufferStream {
-
-		Vector images;
-		int width, height;
-		VideoFormat format;
-
-		int nextImage = 0; // index of the next image to be read.
-		boolean ended = false;
-
-		public ImageSourceStream(int width, int height, int frameRate, Vector images) {
-			this.width = width;
-			this.height = height;
-			this.images = images;
-
-			format = new VideoFormat(VideoFormat.JPEG, new Dimension(width, height), Format.NOT_SPECIFIED,
-					Format.byteArray, (float) frameRate);
-		}
-
-		/**
-		 * We should never need to block assuming data are read from files.
-		 */
-		public boolean willReadBlock() {
-			return false;
-		}
-
-		/**
-		 * This is called from the Processor to read a frame worth of video
-		 * data.
-		 */
-		public void read(Buffer buf) throws IOException {
-
-			// Check if we've finished all the frames.
-			if (nextImage >= images.size()) {
-				// We are done. Set EndOfMedia.
-				System.err.println("Done reading all images.");
-				buf.setEOM(true);
-				buf.setOffset(0);
-				buf.setLength(0);
-				ended = true;
-				return;
-			}
-
-			String imageFile = (String) images.elementAt(nextImage);
-			nextImage++;
-
-			System.err.println("  - reading image file: " + imageFile);
-
-			// Open a random access file for the next image.
-			RandomAccessFile raFile;
-			raFile = new RandomAccessFile(imageFile, "r");
-
-			byte data[] = null;
-
-			// Check the input buffer type & size.
-
-			if (buf.getData() instanceof byte[])
-				data = (byte[]) buf.getData();
-
-			// Check to see the given buffer is big enough for the frame.
-			if (data == null || data.length < raFile.length()) {
-				data = new byte[(int) raFile.length()];
-				buf.setData(data);
-			}
-
-			// Read the entire JPEG image from the file.
-			raFile.readFully(data, 0, (int) raFile.length());
-
-			System.err.println("    read " + raFile.length() + " bytes.");
-
-			buf.setOffset(0);
-			buf.setLength((int) raFile.length());
-			buf.setFormat(format);
-			buf.setFlags(buf.getFlags() | buf.FLAG_KEY_FRAME);
-
-			// Close the random access file.
-			raFile.close();
-		}
-
-		/**
-		 * Return the format of each video frame. That will be JPEG.
-		 */
-		public Format getFormat() {
-			return format;
-		}
-
-		public ContentDescriptor getContentDescriptor() {
-			return new ContentDescriptor(ContentDescriptor.RAW);
-		}
-
-		public long getContentLength() {
-			return 0;
-		}
-
-		public boolean endOfStream() {
-			return ended;
-		}
-
-		public Object[] getControls() {
-			return new Object[0];
-		}
-
-		public Object getControl(String type) {
-			return null;
-		}
-	}
 }
